First pass at moving views into subdir.
