+++ backend/audit/viewlib/__init__.py
@@ -1,0 +1,5 @@
+from .upload_report_view import UploadReportView
+
+views = [
+    UploadReportView,
+]
